DS/emergency-classification/pyproject.toml
[tool.poetry]
name = "emergency-classification"
version = "0.1.0"
description = ""
authors = ["Gutto <guttolaudie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
duckdb = "^0.10.1"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
isort = "^5.13.2"
taskipy = "^1.12.2"
ignr = "^2.2"
black = "^24.3.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.15"
mkdocstrings = "^0.24.1"
mkdocstrings-python = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
